Komponente erstellt
Eine Komponente in Zwei geteilt und Dropdownliste funktioniert einwandsfrei
