+++ src/css/main.css
@@ -1,0 +1,3 @@
+body{
+    background-color: #5050DB;
+}
